pyproject.toml
[tool.poetry]
name = "rssfetcher"
version = "0.1.0"
description = ""
authors = ["Cologler <skyoflw@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
cachetools = "^6.1.0"
fastapi = {extras = ["standard"], version = "^0.116.1"}
pydantic-settings = "^2.6.1"
PyYAML = "^6.0"
requests = "^2.28.1"
schedule = "^1.1.0"


[tool.poetry.group.test.dependencies]
pytest = "^7.1.3"
pytest-asyncio = "^0.19.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
extend-select = [
    "I", # isort
    "ANN", # flake8-annotations
]
